Added triggerOnUpdate
