+++ functions/src/index.ts
@@ -4,7 +4,7 @@ import * as createOrderPackage from "./functions/createOrderFunction";
 import * as sendNotificationPackage from "./functions/sendNotificationFunction";
 import * as startTripPackage from "./functions/startTripFunction";
 import * as finishTripPackage from "./functions/finishTripFunction";
-import {sendNotificationMessage} from "./common/functions/sendNotificationMessage";
+import * as triggers from "./triggers";
 
 admin.initializeApp();
 
@@ -56,44 +56,8 @@ export const finishTrip = functions.https.onRequest((request, response) => {
 // // https://firebase.google.com/docs/functions/firestore-events
 //
 
-export const usersTrigger = functions.firestore.document("c_users/{id}").onUpdate((change, context) => {
-  const firebaseFirestore = admin.firestore();
-  const sessionsCollection = firebaseFirestore.collection("c_sessions");
-
-  const after = change.after.data();
-  const userId = after.id;
-  const userName = after.name;
-
-  functions.logger.info(`User identified ${userId}`);
-
-  if (after.is_driver === "1" && userName != null) {
-    functions.logger.info(`User is verified driver ${after.is_driver}`);
-
-    sessionsCollection.where("id_users", "==", `${userId}`).get().then((querySnapshot) => {
-      const tokenMessaging: string[] = [];
-      functions.logger.info(`Sessions filtered count ${querySnapshot.size}`);
-      querySnapshot.forEach((doc) => {
-        const sessionData = doc.data();
-        if (sessionData.token_messaging != null) tokenMessaging.push(sessionData.token_messaging);
-        sessionsCollection.doc(`${sessionData.id_sessions}`).update({is_driver: true}).then(() => {
-          functions.logger.info(`Session Updated ${sessionData.id_sessions}`);
-        }).catch((error: any) => {
-          functions.logger.error("Error updating session", error);
-        });
-      });
-      functions.logger.info(`Sending message to ${tokenMessaging.length} tokens`);
-      sendNotificationMessage(tokenMessaging, {
-        notification: {
-          title: "¡HORA DE ROQUEAR!",
-          body: `Bienvenido, ya eres CONDUCTOR!, ${userName}`,
-        },
-      }).then((response) => {
-        functions.logger.info(`Messages sends response ${response}`);
-      }).catch((error: any) => {
-        functions.logger.error("Error sending messages", error);
-      });
-    }).catch((error: any) => {
-      functions.logger.error("Error where id_users", error);
-    });
-  }
-});
+export const triggerOnUpdate = functions.firestore.document("{collection}/{id}").onUpdate((change, context) =>{
+  functions.logger.info("TRIGGER onUpdate");
+  triggers.handler(change, context);
+}
+);
