+++ functions/src/triggers/users/index.ts
@@ -1,0 +1,5 @@
+import {handler as onUpdate} from "./update";
+
+export {
+  onUpdate,
+};
